Support postgres:12.2 with auth 'trust'
Behavior changed in postgres:12.2, now it is required to either set up
a password for the postgres superuser or to force the auth method to be
'trust'. Since we're running this in a CI and databases are volatile
here either way I chose the 'trust' version.

See https://github.com/docker-library/postgres/commit/42ce7437ee68150ee29f5272428aa4fc657dc6dc
for the commit on the postgres image side.